team_sa_ayd/app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq>=0.4.0
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
